Modularize application builder
