+++ src/app/mod.rs
@@ -1,0 +1,10 @@
+//! Application builder and related types.
+
+mod builder;
+mod connection;
+mod envelope;
+mod error;
+
+pub use builder::{ConnectionSetup, ConnectionTeardown, Handler, Middleware, WireframeApp};
+pub use envelope::{Envelope, Packet, PacketParts};
+pub use error::{Result, SendError, WireframeError};
